Integrate SDL3 library into RayTracer project
Added `sdl3.lib` to the `<AdditionalDependencies>` property in
`RayTracer.vcxproj` for all configurations and platforms to
ensure proper linking of the SDL3 library. Updated the
`<AdditionalLibraryDirectories>` property to include the path
to the SDL3 library directory
(`$(SolutionDir)ThirdParty\SDL3\lib\$(PlatformTarget)`) for
all configurations and platforms. These changes enable the
project to compile and link successfully with SDL3.